Separate sequencer and full node e2e tests
